New custom localization mechanism, that allows for easy adding of new items and safe and straightforward use in various forms, thanks to the use of internal dictionary. Some parts of code were reorganized and all code related to updater mechanism was moved under Form_Updates class. Modified icons for Updater and About dialog. Several smaller code improvements. Added comments to the code. This is a candidate for 1.1.0 release.

+++ Morseapp_WinForms/Classes/Morse.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
@@ -241,7 +240,6 @@ namespace Morseapp_WinForms
             return "Success.";
         }
 
-
         /// <summary>
         /// Method that returns morseList as table-formatted string.
         /// </summary>
@@ -249,7 +247,7 @@ namespace Morseapp_WinForms
         {
             string table = "";
             table += $"ASCII\tMorse\t\tASCII\tMorse\t\tASCII\tMorse\t\tASCII\tMorse{Environment.NewLine}";
-            ushort counter = 0;
+            byte counter = 0;
 
             foreach (var item in morseList)
             {
